+++ engine/deskelectronjs/src/reaction.js
@@ -109,7 +109,7 @@ module.exports = async (...args) => {
             // let iscss = handler.check_empty(options.css);
 
             if (!isredirect) {
-              res.redirect(301, redirect);
+              res.redirect(redirect);
               resolve(rtn);
             } else if (!isjson) {
               res.status(status).json(json);
@@ -193,6 +193,152 @@ module.exports = async (...args) => {
                   collapseWhitespace: true,
                 })
               );
+            } else {
+              throw {
+                message:
+                  "Cannot found any html file name from view or options.layout property or html in string type for render!",
+                stack:
+                  "Cannot found any html file name from view or options.layout property or html in string type for render!",
+              };
+            }
+
+            resolve(rtn);
+          } catch (error) {
+            if (error.errno)
+              resolve({
+                code: error.errno,
+                errno: error.errno,
+                message: error.message,
+                stack: error.stack,
+                data: null,
+              });
+            else {
+              let err = {
+                code: 506,
+                errno: 506,
+                message: "",
+                stack: "",
+                data: null,
+              };
+              if (error.msg) {
+                err.message = error.msg;
+                err.stack = error.data;
+              } else {
+                err.message = error.message;
+                err.stack = error.stack;
+              }
+              resolve(err);
+            }
+          }
+        });
+      };
+
+      /**
+       * The final process redirect gui content to frontend
+       * @alias module:reaction.procredirect
+       * @param {...Object} args - 1 parameters
+       * @param {Object} args[0] - res the object for render to frontend
+       */
+      const procredirect = (...args) => {
+        return new Promise(async (resolve, reject) => {
+          const { JSDOM } = jsdom;
+          let [res] = args;
+          try {
+            let {
+              options: {
+                css,
+                html,
+                injectionjs,
+                js,
+                json,
+                layer,
+                less,
+                mjs,
+                params,
+                redirect,
+              },
+              status,
+              view,
+            } = res.locals.render;
+            let rtn = handler.dataformat;
+            let isview = handler.check_empty(view);
+            let islayer = handler.check_empty(layer.layouts);
+            let ishtml = handler.check_empty(html);
+            // let iscss = handler.check_empty(options.css);
+
+            if (!islayer || !isview || !ishtml) {
+              let dom, isvalid, layouts;
+              if (!ishtml) dom = new JSDOM(html);
+              else {
+                if (!isview) {
+                  if (path.extname(view) == ".html") isvalid = true;
+                  else isvalid = molecule.indentify_html(view);
+                } else if (!islayer) {
+                  if (path.extname(layer.layouts) == ".html") isvalid = true;
+                  else isvalid = molecule.indentify_html(layer.layouts);
+                }
+                if (isvalid) {
+                  if (!islayer) {
+                    if (!isview) layer.childs.external.push(view);
+                    let { code, msg, data } = await molecule.combine_layer(
+                      layer,
+                      params
+                    );
+                    if (code == 0) layouts = data;
+                    else throw { msg: msg, data: data };
+                  } else if (!isview) {
+                    let { code, msg, data } = await molecule.single_layer(
+                      view,
+                      params
+                    );
+
+                    if (code == 0) view = data;
+                    else throw { msg: msg, data: data };
+                  }
+
+                  if (!layouts) {
+                    dom = new JSDOM(view);
+                  } else {
+                    dom = new JSDOM(layouts);
+                  }
+                } else {
+                  throw {
+                    message:
+                      "Cannot found any html extension file from view or options.layout property!",
+                    stack:
+                      "Cannot found any html extension file from view or options.layout property!",
+                  };
+                }
+              }
+              let document = dom.window.document;
+              for (let [el, content] of Object.entries(params)) {
+                let found = document.querySelector(el);
+                if (found) found.innerHTML = content;
+              }
+
+              let preload = await molecule.get_domhtml(
+                path.join(pathname, "browser", "preload.html")
+              );
+              document.querySelector("body").innerHTML += preload;
+              let script = document.createElement("script");
+              script.type = "text/javascript";
+              script.innerHTML = `var mjs=${JSON.stringify(
+                molecule.import_mjs(mjs, params)
+              )};`;
+              if (Object.keys(injectionjs.variables).length > 0)
+                script.innerHTML += `var injectionjs=${JSON.stringify(
+                  injectionjs.variables
+                )}`;
+              document.getElementsByTagName("head")[0].appendChild(script);
+              let rtnimport_css = molecule.import_css(document, css, params);
+              if (rtnimport_css) throw rtnimport_css;
+              let rtnimport_js = molecule.import_js(document, js, params);
+              if (rtnimport_js) throw rtnimport_js;
+              let rtnimport_less = molecule.import_less(document, less, params);
+              if (rtnimport_less) throw rtnimport_less;
+              rtn.data = await minify(dom.serialize(), {
+                collapseWhitespace: true,
+              });
             } else {
               throw {
                 message:
@@ -600,6 +746,211 @@ module.exports = async (...args) => {
         }
       };
 
+      /**
+       * The main objective is on listen register url from redirectly gui
+       * @alias module:reaction.onredirect
+       * @param {...Object} args - 2 parameters
+       * @param {Array} args[0] - orireq http request module
+       * @param {Array} args[1] - orires http response module
+       */
+      lib["onredirect"] = async (...args) => {
+        let [orireq, orires, next] = args;
+        let fn;
+        try {
+          //Resolve web page caching across all browsers
+          //https://stackoverflow.com/questions/49547/how-do-we-control-web-page-caching-across-all-browsers
+          orires.setHeader(
+            "Cache-Control",
+            "no-cache, no-store, must-revalidate"
+          ); // HTTP 1.1.
+          orires.setHeader("Pragma", "no-cache"); // HTTP 1.0.
+          orires.setHeader("Expires", "0"); // Proxies.
+
+          let paramres = {};
+          let paramerror;
+          let errorstack;
+          let redirect;
+          let { defaulturl, ...component } = components;
+          for (let [compkey, compval] of Object.entries(component)) {
+            let { api, gui } = compval;
+            let baseUrl = orireq.originalUrl;
+            if (!handler.check_empty(orireq?.params)) {
+              fn = await guiapi_route_filter(api, gui, {
+                params: orireq.params,
+                originalUrl: orireq.originalUrl,
+              });
+            } else {
+              let pos = orireq.originalUrl.indexOf("?");
+              if (pos != -1) baseUrl = orireq.originalUrl.substring(0, pos);
+              fn = await guiapi_picker(getNestedObject(gui, baseUrl));
+            }
+            if (fn) {
+              orireq.method = fn.method;
+              break;
+            }
+          }
+
+          if (fn) {
+            if (orireq.method == fn.method.toUpperCase()) {
+              let permit = true;
+              for (let [idx, queue] of fn["controller"].entries()) {
+                let fname = Object.keys(queue)[0];
+                paramres["onreq"] = {
+                  fname: fname,
+                  index: idx,
+                };
+                let {
+                  err: chkerr,
+                  fname: chkfname,
+                  render: chkrender,
+                  rule: chkrule,
+                  ...chkparamres
+                } = paramres;
+
+                let response = {
+                  err: {
+                    error: "",
+                    render: handler.webview,
+                  },
+                  fname: fname,
+                  render: handler.webview,
+                  rule: {},
+                  ...chkparamres,
+                };
+                if (chkrule) response = { ...response, rule: chkrule };
+
+                let queuertn;
+                if (permit && fn.idx == idx) {
+                  queuertn = sanbox(queue[fname], [orireq, response]);
+                  if (queuertn instanceof Promise) queuertn = await queuertn;
+                  queuertn["action"] = queuertn.render;
+                } else if (fn.idx != idx) {
+                  queuertn = sanbox(queue[fname], [orireq, response]);
+                  if (queuertn instanceof Promise) queuertn = await queuertn;
+                  queuertn["action"] = queuertn.render;
+                }
+
+                let { err, render, stack, message, ...res } = queuertn;
+                if (stack && message) {
+                  if (!err)
+                    err = {
+                      error: stack,
+                      render: handler.webview,
+                      ...res.data,
+                    };
+                  if (!res.action) res.action = err.render;
+                }
+                if (!res.action) {
+                  delete queuertn["action"];
+                  throw queuertn;
+                }
+
+                if (
+                  err &&
+                  (!isrender(err.render) || !handler.check_empty(err.error))
+                ) {
+                  if (
+                    permit &&
+                    idx >= fn.idx &&
+                    idx + 1 <= fn.controller.length
+                  ) {
+                    if (errorstack) errorstack = { ...paramerror, ...err };
+                    else errorstack = { ...{}, ...paramerror, ...err };
+                  } else {
+                    paramerror = { ...paramerror, ...err };
+                    if (!isrender(err.render) && permit == true) permit = false;
+                  }
+                }
+
+                if (render) {
+                  if (!isrender(render)) paramres["render"] = render;
+                }
+                if (res) {
+                  paramres = { ...paramres, ...res };
+                  if (
+                    permit &&
+                    idx < fn.idx &&
+                    res.action.options.redirect != ""
+                  )
+                    break;
+                  if (!permit && idx > fn.idx && !fn.strict) break;
+                }
+                if (paramerror !== undefined) break;
+              }
+            } else {
+              paramerror = {
+                render: handler,
+                error: "Unmatched the request method!",
+              };
+            }
+
+            if (errorstack) paramerror = errorstack;
+            // Error checking
+            if (paramerror) {
+              orires.locals = {
+                render: handler.webview,
+              };
+              if (!handler.check_empty(paramerror.error))
+                throw { code: 500, message: paramerror.error };
+              if (!isrender(paramerror.render))
+                orires.locals.render = paramerror.render;
+            } else if (paramres) {
+              if (paramres.render) {
+                if (!isrender(paramres.render))
+                  orires.locals = { render: paramres.render };
+              } else orires.locals = { render: handler.webview };
+            }
+          }
+
+          let rtn = await procredirect(orires);
+
+          if (rtn.code !== 0) throw rtn;
+          return rtn;
+        } catch (error) {
+          orires.locals = { render: handler.webview };
+          let { render: err } = orires.locals;
+          let errcode, errmessage;
+
+          if (error.code) {
+            errcode = error.code;
+            errmessage = error.message;
+          } else if (error.err) {
+            errcode = 506;
+            if (error.err.error == "") errmessage = "Unexpected error!";
+            else errmessage = error.err.error;
+          } else {
+            errcode = 506;
+            errmessage = error;
+          }
+          if (fn?.from == "api") {
+            err["options"]["json"] = {
+              ...handler.dataformat,
+              ...{ code: errcode, msg: errmessage },
+            };
+          } else {
+            err["status"] = errcode;
+            if (errcode >= 500)
+              err["view"] = `${pathname}/browser/error/500.html`;
+            else err["view"] = `${pathname}/browser/error/404.html`;
+            let msg;
+            if (typeof errmessage == "string") msg = errmessage;
+            else msg = JSON.stringify(errmessage);
+
+            err["options"]["params"] = {
+              errorcode: errcode,
+              title: "System Notification",
+              msg: msg,
+            };
+          }
+          let errmsg = `"${orireq.method} ${orireq.originalUrl} HTTP/1.1" ${errcode} Error:`;
+          if (typeof errmessage == "string") errmsg += errmessage;
+          else errmsg += JSON.stringify(errmessage);
+          logerr(errmsg);
+
+          return await procredirect(orires);
+        }
+      };
+
       lib["onless"] = async (...args) => {
         let [orireq, orires] = args;
         let fn;
@@ -617,7 +968,6 @@ module.exports = async (...args) => {
           }
         } catch (error) {
         } finally {
-          orires.end();
         }
       };